refactor: 로또 게임을 위한 준비 로직 Controller로 분리
- 로또 금액 입력 및 저장
- 사용자 로또 발행

- 당첨 로또 번호 발행
- 당첨 로또 보너스 번호 발행
